backend/foodgram/api/filters.py
from django_filters import rest_framework as filters
from recipes.models import Recipe


class RecipeFilters(filters.FilterSet):
    tags = filters.CharFilter(field_name='tags__slug',
                              lookup_expr='iexact')

    is_favorited = filters.BooleanFilter(
        method='get_is_favorited'
    )
    is_in_shopping_cart = filters.BooleanFilter(
        method='get_is_in_shopping_cart'
    )

    class Meta:
        model = Recipe
        fields = ('tags', 'author', 'is_favorited', 'is_in_shopping_cart')

    def get_is_favorited(self, queryset, name, value):
        user = self.request.user
        if value:
            return queryset.filter(favorited__user=user)
        return Recipe.objects.all()

    def get_is_in_shopping_cart(self, queryset, name, value):
        user = self.request.user
        if value:
            return queryset.filter(recipe_cart__user=user)
        return Recipe.objects.all()
