requirements.txt
asgiref==3.5.2
astroid==2.12.10
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.5.1
Django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==5.0.4
idna==3.4
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.5.0
pylint==2.15.3
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.5
typed-ast==1.5.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
zipp==3.8.1